Changed Fee to Big types.

+++ src/main/java/io/github/ngchinhow/kraken/rest/model/marketdata/pair/Fee.java
@@ -4,10 +4,13 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Data
 @JsonFormat(shape = JsonFormat.Shape.ARRAY)
 @JsonPropertyOrder({"minVolume", "percentFee"})
 public class Fee {
-    private int minVolume;
-    private int percentFee;
+    private BigInteger minVolume;
+    private BigDecimal percentFee;
 }
